refactor: Create workspace
So that I can have a bin that pulls from a lib, and eventually even a
GUI app that pulls from the lib too.
