+++ gdal-dtm-exporter/Cargo.toml
@@ -1,14 +1,16 @@
-[package]
-name = "gdal-dtm-exporter"
-version = "0.1.0"
+[workspace]
+resolver = "2"
+members = [
+    "gdal-dtm-exporter-bin",
+    "gdal-dtm-exporter-lib",
+    "gdal-dtm-exporter-gui",
+]
+
+default-members = [
+    "gdal-dtm-exporter-bin",
+    "gdal-dtm-exporter-lib",
+    "gdal-dtm-exporter-gui",
+]
+
+[workspace.package]
 edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap = { version = "4.5.4", features = ["derive"] }
-color-eyre = "0.6.3"
-env_logger = "0.11.3"
-gdal = { version = "0.16.0", features = ["bindgen"] }
-image = "0.25.1"
-log = "0.4.21"
