+++ gdal-dtm-exporter/gdal-dtm-exporter-bin/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "gdal-dtm-exporter-bin"
+version = "0.1.0"
+edition = "2021"
+
+[[bin]]
+name = "gdal-dtm-exporter"
+path = "src/main.rs"
+
+[dependencies]
+gdal-dtm-exporter-lib = { path = "../gdal-dtm-exporter-lib" }
+
+clap = { version = "4.5.4", features = ["derive"] }
+color-eyre = "0.6.3"
+env_logger = "0.11.3"
+gdal = { version = "0.16.0", features = ["bindgen"] }
+image = "0.25.1"
+log = "0.4.21"
