+++ gdal-dtm-exporter/gdal-dtm-exporter-lib/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "gdal-dtm-exporter-lib"
+version = "0.1.0"
+edition = "2021"
+description = "Library to export OpenEXR images from DTM elevation files."
+
+[lib]
+name = "gdal_dtm_exporter_lib"
+path = "src/mod.rs"
+
+[dependencies]
+clap = { version = "4.5.4", features = ["derive"] }
+color-eyre = "0.6.3"
+env_logger = "0.11.3"
+gdal = { version = "0.16.0", features = ["bindgen"] }
+image = "0.25.1"
+log = "0.4.21"
